Restructure this repository to have a more conventional project layout.
This is done by creating directories for notifier code and test code.
Moving those files into separate directories required some other
changes. The modules names needed to change in the code to reflect the
new structure. Pytest.ini file was created to implement pythonpath for
pytest. Containerfile modified as gunicorn cmd needed to macth new
structure.
