+++ Containerfile
@@ -6,4 +6,4 @@ RUN ansible-galaxy collection install dellemc.os9 && \
 	ansible-galaxy collection install amazon.aws
 COPY . ./
 
-CMD [ "gunicorn", "-b", "0.0.0.0:8000", "app:app"]
+CMD [ "gunicorn", "-b", "0.0.0.0:8000", "slack_notifier.app:app"]
